refactor: Move application models into their own files
